+++ pdi/include/pdi/context.h
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (C) 2015-2024 Commissariat a l'energie atomique et aux energies alternatives (CEA)
+ * Copyright (C) 2015-2026 Commissariat a l'energie atomique et aux energies alternatives (CEA)
  * Copyright (C) 2021 Institute of Bioorganic Chemistry Polish Academy of Science (PSNC)
  * All rights reserved.
  *
@@ -135,6 +135,23 @@ public:
 
 	/// Finalizes PDI and exits application
 	virtual void finalize_and_exit() = 0;
+};
+
+class PDI_EXPORT TimerEventHandler
+{
+private:
+	Context& m_ctx;
+	const std::string m_plugin_name;
+
+public:
+	TimerEventHandler(Context& ctx, const std::string& plugin_name)
+		: m_ctx(ctx)
+		, m_plugin_name(plugin_name)
+	{
+		m_ctx.event((plugin_name + "_start_timer").c_str());
+	}
+
+	~TimerEventHandler() { m_ctx.event((m_plugin_name + "_stop_timer").c_str()); }
 };
 
 } // namespace PDI