+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to
 ### For users
 
 #### Added
+* A timer plugin is added to reflect the time consumed by pdi and its plugins.
+  [#710](https://github.com/pdidev/pdi/issues/710)
 * Add test for HDF5 precision conversion [#708](https://github.com/pdidev/pdi/issues/708)
 * Improved messages for specification tree errors, with file & line numbers and
   support for file names from Paraconf 1.1,